Use bnum of num bigint
This patch represents a major overhaul of this crate, moving it from a
wrapper around num_bigint to a crate called bnum which uses const
generics to achieve clone-able integers of dynamic size at compile time
but fixed size at runtime.

The goal here is both to reduce the number of dependencies involved in
this crate, increase the number of native passthrough operators, and to
eliminate the constant need for clone while using this crate.