Fix integer size
This patch fixes an integer size problem, as it turns out the const
generic factor for integer width in bnum is in units of 64bits, not
single bits. Meaning this library was using 16,384 bit integers not 256
bit! Since the internal type is exposed publicly this is a breaking
change and requires a version bump.